move `DistMult` to `kale.predict.devode`
